pyproject.toml
[project]
name = "iris-oc-mapper"
version = "1.0.1"
description = "A tool for mapping CINECA IRIS bibliographic records to OpenCitations Meta and Index datasets, with built-in utilities for interacting with IRIS data dumps."
authors = [{ name = "Leonardo Zilli", email = "leonardo.zilli@studio.unibo.it" }]
readme = "README.md"
license = { file = "LICENSE" }
keywords = ["IRIS", "OpenCitations", "Mapping", "Bibliometrics", "Research Information Systems"]
classifiers = [
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Intended Audience :: Science/Research",
]
requires-python = ">=3.12"

dependencies = [
    "isal>=1.8.0",
    "jinja2>=3.1.6",
    "pandera[polars]>=0.26.1",
    "plotly>=6.3.0",
    "polars>=1.31.0",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.9",
    "pyyaml>=6.0.2",
    "typer>=0.17.4",
]

[dependency-groups]
dev = [
    "distinctipy>=1.3.4",
    "ipykernel>=6.29.5",
    "tabulate>=0.9.0",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-mock>=3.15.0",
    "profilehooks>=1.13.0",
    "snakeviz>=2.2.2",
    "fastexcel>=0.16.0",
    "nbformat>=5.10.4",
    "twine>=6.2.0",
    "mypy>=1.18.2",
    "pyarrow-stubs>=20.0.0.20251107",
]

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[project.scripts]
iris-oc-mapper = "iris_oc_mapper.cli:app"

[tool.mypy]
disable_error_code = ["import-untyped"]

[tool.bumpver]
current_version = "1.0.2"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
tag_message = "{new_version}"
tag_scope = "default"
pre_commit_hook = ""
post_commit_hook = ""
commit =false
tag = false
push = false

[tool.bumpver.file_patterns]
"pyproject.toml" = ["version = \"{version}\""]
"src/iris_oc_mapper/__init__.py" = [
    '__version__ = "{version}"',
]

